status now standalone
